to_uint now returns a uint64_t

+++ src/common/Variant.h
@@ -118,8 +118,8 @@ public:
     cali_id_t      to_id(bool* okptr = nullptr) const;
     int            to_int(bool* okptr = nullptr);
     int            to_int(bool* okptr = nullptr) const;
-    unsigned       to_uint(bool* okptr = nullptr);
-    unsigned       to_uint(bool* okptr = nullptr) const;
+    uint64_t       to_uint(bool* okptr = nullptr);
+    uint64_t       to_uint(bool* okptr = nullptr) const;
     bool           to_bool(bool* okptr = nullptr);
     double         to_double(bool* okptr = nullptr);
     double         to_double(bool* okptr = nullptr) const;
